refactor: Moved dataclasses to types.py
